requirements.txt
boto3==1.39.11
botocore==1.39.11
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
jmespath==1.0.1
python-dateutil==2.9.0.post0
requests==2.32.4
s3transfer==0.13.1
six==1.17.0
urllib3==2.5.0
pydantic==2.10.3
PyYAML==6.0.1

# Azure SDK dependencies
azure-identity==1.15.0
azure-mgmt-compute==30.4.0
azure-mgmt-storage==21.1.0
azure-mgmt-network==25.1.0
azure-mgmt-keyvault==10.3.0
azure-mgmt-authorization==4.0.0
azure-mgmt-msi==7.0.0
azure-mgmt-security==6.0.0
azure-mgmt-resource==23.0.1
azure-mgmt-subscription==3.1.1

# FULL SET
annotated-types==0.7.0
asttokens==3.0.0
azure-common==1.1.28
azure-core==1.35.0
azure-identity==1.15.0
azure-mgmt-authorization==4.0.0
azure-mgmt-compute==30.4.0
azure-mgmt-core==1.6.0
azure-mgmt-keyvault==10.3.0
azure-mgmt-msi==7.0.0
azure-mgmt-network==25.1.0
azure-mgmt-resource==23.0.1
azure-mgmt-security==6.0.0
azure-mgmt-storage==21.1.0
azure-mgmt-subscription==3.1.1
boto3==1.39.11
botocore==1.39.11
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.5
decorator==5.2.1
executing==2.2.0
idna==3.10
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
jmespath==1.0.1
matplotlib-inline==0.1.7
msal==1.33.0
msal-extensions==1.3.1
msrest==0.7.1
oauthlib==3.3.1
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyGithub==2.7.0
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.1
requests==2.32.4
requests-oauthlib==2.0.0
s3transfer==0.13.1
six==1.17.0
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uv==0.8.4
wcwidth==0.2.13
